requirements.txt
opencv-python==4.10.0
openpyxl==3.1.5
Pillow==10.4.0
ultralytics==8.2.76